Move sensors to Draggable
