+++ src/Draggable/Sensors/ForceTouchSensor/ForceTouchSensor.js
