+++ src/Draggable/Sensors/ForceTouchSensor/index.js
