+++ src/Draggable/Sensors/Sensor/Sensor.js
