+++ src/Draggable/Sensors/SensorEvent/SensorEvent.js
